fix/support-decimal-values: change parse int to float so upport floats

+++ src/components/rpe/RentalPropertyEvaluator.js
@@ -16,7 +16,7 @@ class RentalPropertyEvaluator extends React.Component {
 		console.log('field has changed ', inputChanged, newValue)
 		await this.setState( ( prevState ) => {
 			const newState = { ...prevState };
-			newState.changeable[inputChanged] = parseInt( newValue );
+			newState.changeable[inputChanged] = parseFloat( newValue );
 			return newState;
 		})
 		await this.calcAllDynamically();
